backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Flask==2.3.2
Flask-Cors==4.0.0
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.25.2
openai==0.27.8
packaging==23.1
pandas==2.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
threadpoolctl==3.2.0
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.3.6
xgboost==1.7.6
yarl==1.9.2
zipp==3.16.2
